+++ src/Farmer/Arm/AVS.fs
@@ -1,0 +1,5 @@
+﻿[<AutoOpen>]
+module Farmer.Arm.AVS
+
+open Farmer
+let privateClouds = ResourceType("Microsoft.AVS/privateClouds", "2021-12-01")