+++ src/Farmer/Arm/Network.fs
@@ -3,7 +3,9 @@ module Farmer.Arm.Network
 
 open System.Net.Mail
 open Farmer
+open Farmer.Arm
 open Farmer.ExpressRoute
+open Farmer.Route
 open Farmer.VirtualNetworkGateway
 
 let connections = ResourceType("Microsoft.Network/connections", "2020-04-01")
@@ -23,6 +25,9 @@ let networkProfiles =
 let publicIPAddresses =
     ResourceType("Microsoft.Network/publicIPAddresses", "2018-11-01")
 
+let publicIPPrefixes =
+    ResourceType("Microsoft.Network/publicIPPrefixes", "2021-08-01")
+
 let serviceEndpointPolicies =
     ResourceType("Microsoft.Network/serviceEndpointPolicies", "2020-07-01")
 
@@ -38,11 +43,17 @@ let virtualNetworkGateways =
 let localNetworkGateways =
     ResourceType("Microsoft.Network/localNetworkGateways", "")
 
+let natGateways = ResourceType("Microsoft.Network/natGateways", "2021-08-01")
+
 let privateEndpoints =
     ResourceType("Microsoft.Network/privateEndpoints", "2021-05-01")
 
 let virtualNetworkPeering =
     ResourceType("Microsoft.Network/virtualNetworks/virtualNetworkPeerings", "2020-05-01")
+
+let routeTables = ResourceType("Microsoft.Network/routeTables", "2021-01-01")
+let routes = ResourceType("Microsoft.Network/routeTables/routes", "2021-01-01")
+
 
 type SubnetReference =
     | ViaManagedVNet of (ResourceId * ResourceName)
@@ -84,6 +95,59 @@ type SubnetReference =
 
         Direct subnetRef
 
+type Route =
+    {
+        Name: ResourceName
+        AddressPrefix: IPAddressCidr
+        NextHopType: Route.HopType
+        HasBgpOverride: FeatureFlag
+    }
+
+    member internal this.JsonModelProperties =
+        {|
+            addressPrefix = IPAddressCidr.format this.AddressPrefix
+            nextHopType = this.NextHopType.ArmValue
+            nextHopIpAddress =
+                match this.NextHopType with
+                | VirtualAppliance ip ->
+                    ip
+                    |> Option.map (fun x -> x.ToString())
+                    |> Option.defaultValue Unchecked.defaultof<_>
+                | _ -> Unchecked.defaultof<_>
+            hasBgpOverride = this.HasBgpOverride.AsBoolean
+        |}
+
+    interface IArmResource with
+        member this.ResourceId = routes.resourceId this.Name
+
+        member this.JsonModel =
+            {| routes.Create(this.Name) with
+                properties = this.JsonModelProperties
+            |}
+
+type RouteTable =
+    {
+        Name: ResourceName
+        Location: Location
+        Tags: Map<string, string>
+        DisableBGPRoutePropagation: FeatureFlag
+        Routes: Route list
+    }
+
+    member internal this.JsonModelProperties =
+        {|
+            disableBgpRoutePropagation = this.DisableBGPRoutePropagation.AsBoolean
+            routes = this.Routes |> Seq.map (fun x -> (x :> IArmResource).JsonModel)
+        |}
+
+    interface IArmResource with
+        member this.ResourceId = routeTables.resourceId this.Name
+
+        member this.JsonModel =
+            {| routeTables.Create(this.Name, this.Location, tags = this.Tags) with
+                properties = this.JsonModelProperties
+            |}
+
 type PublicIpAddress =
     {
         Name: ResourceName
@@ -110,6 +174,44 @@ type PublicIpAddress =
                     |}
             |}
 
+/// If using the IPs in the frontend of a cross-region laod balancer, public IPs and prefixes must be in
+/// the Global tier, otherwise regional IPs are sufficient.
+type PublicIpPrefixTier =
+    | Global
+    | Regional
+
+    member this.ArmValue =
+        match this with
+        | Global -> "Global"
+        | Regional -> "Regional"
+
+/// Public IP Prefix creates a block of contiguous public IP addresses that can be assigned to resources.
+type PublicIpPrefix =
+    {
+        Name: ResourceName
+        Location: Location
+        PrefixLength: int
+        Tier: PublicIpPrefixTier
+        Tags: Map<string, string>
+    }
+
+    interface IArmResource with
+        member this.ResourceId = publicIPPrefixes.resourceId this.Name
+
+        member this.JsonModel =
+            {| publicIPPrefixes.Create(this.Name, this.Location, tags = this.Tags) with
+                sku =
+                    {|
+                        name = "Standard"
+                        tier = this.Tier.ArmValue
+                    |}
+                properties =
+                    {|
+                        prefixLength = this.PrefixLength
+                        publicIPAddressVersion = "IPv4"
+                    |}
+            |}
+
 type SubnetDelegation =
     {
         Name: ResourceName
@@ -123,6 +225,7 @@ type Subnet =
         VirtualNetwork: LinkedResource option
         NetworkSecurityGroup: LinkedResource option
         Delegations: SubnetDelegation list
+        NatGateway: LinkedResource option
         ServiceEndpoints: (Network.EndpointServiceType * Location list) list
         AssociatedServiceEndpointPolicies: ResourceId list
         PrivateEndpointNetworkPolicies: FeatureFlag option
@@ -132,6 +235,10 @@ type Subnet =
     member internal this.JsonModelProperties =
         {|
             addressPrefix = this.Prefix
+            natGateway =
+                this.NatGateway
+                |> Option.map LinkedResource.AsIdObject
+                |> Option.defaultValue Unchecked.defaultof<_>
             networkSecurityGroup =
                 this.NetworkSecurityGroup
                 |> Option.map (fun nsg ->
@@ -211,6 +318,10 @@ type VirtualNetwork =
                 seq {
                     for subnet in this.Subnets do
                         match subnet.NetworkSecurityGroup with
+                        | Some (Managed id) -> id
+                        | _ -> ()
+
+                        match subnet.NatGateway with
                         | Some (Managed id) -> id
                         | _ -> ()
                 }
@@ -497,8 +608,10 @@ type NetworkInterface =
                                         subnet =
                                             {|
                                                 id =
-                                                    subnets
-                                                        .resourceId(this.VirtualNetwork.Name, ipConfig.SubnetName)
+                                                    { this.VirtualNetwork.ResourceId with
+                                                        Type = subnets
+                                                        Segments = [ ipConfig.SubnetName ]
+                                                    }
                                                         .Eval()
                                             |}
                                     |}
@@ -772,5 +885,38 @@ type NetworkPeering =
                                     | Managed id
                                     | Unmanaged id -> id.ArmExpression.Eval()
                             |}
+                    |}
+            |}
+
+type NatGateway =
+    {
+        Name: ResourceName
+        Location: Location
+        IdleTimeout: int<Minutes>
+        PublicIpAddresses: LinkedResource list
+        PublicIpPrefixes: LinkedResource list
+        Tags: Map<string, string>
+    }
+
+    interface IArmResource with
+        member this.ResourceId = natGateways.resourceId this.Name
+
+        member this.JsonModel =
+            let dependencies =
+                seq {
+                    for linkedResource in this.PublicIpAddresses @ this.PublicIpPrefixes do
+                        match linkedResource with
+                        | Managed resId -> resId
+                        | _ -> ()
+                }
+                |> Set.ofSeq
+
+            {| natGateways.Create(this.Name, this.Location, dependsOn = dependencies, tags = this.Tags) with
+                sku = {| name = "Standard" |}
+                properties =
+                    {|
+                        idleTimeoutInMinutes = this.IdleTimeout
+                        publicIpAddresses = this.PublicIpAddresses |> List.map LinkedResource.AsIdObject
+                        publicIpPrefixes = this.PublicIpPrefixes |> List.map LinkedResource.AsIdObject
                     |}
             |}