+++ src/Tests/ContainerApps.fs
@@ -32,6 +32,7 @@ let fullContainerAppDeployment =
         }
 
     let version = "1.0.0"
+    let managedIdentity = ManagedIdentity.Empty
 
     let containerEnv =
         containerEnvironment {
@@ -46,7 +47,8 @@ let fullContainerAppDeployment =
                         add_identity msi
                         active_revision_mode Single
 
-                        add_registry_credentials [ registry containerRegistryDomain containerRegistryName ]
+                        add_registry_credentials
+                            [ registry containerRegistryDomain containerRegistryName managedIdentity ]
 
                         add_containers
                             [